preprocessing/requirements.txt
sshtunnel
pandas
seaborn
scikit-learn
psycopg2
configparser
requests
lxml
beautifulsoup4
boilerpipe3